slider: move example utils to separate file
